Add Javadoc comments to LLVMFunctionRegistry

+++ dataset/GitHub_Java/oracle.graal/projects/com.oracle.truffle.llvm.nodes.impl/src/com/oracle/truffle/llvm/nodes/impl/base/LLVMFunctionRegistry.java
@@ -150,8 +150,8 @@ public class LLVMFunctionRegistry {
      * @param index the function index
      * @return the function descriptor
      */
-    public LLVMFunctionDescriptor createFromIndex(int index) {
-        LLVMFunctionDescriptor llvmFunction = functionDescriptors[index];
+    public LLVMFunctionDescriptor createFromIndex(long index) {
+        LLVMFunctionDescriptor llvmFunction = functionDescriptors[(int) index];
         assert llvmFunction != null;
         return llvmFunction;
     }
